requirements.txt
# Core Engine Dependencies (Week 3/4)
typer==0.12.3
click==8.1.7
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
rich==13.7.0
pyyaml==6.0.1
pydantic-settings==2.0.0
pathlib

# AI & Vectors
sentence-transformers==2.3.1
numpy==1.26.3

# Document Loaders (MonsterLoader)
pypdf==4.0.1
ebooklib==0.18
beautifulsoup4==4.12.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.2.0

# Images & OCR
Pillow==10.2.0
pytesseract==0.3.10

# Utils
httpx==0.26.0
chardet==5.2.0
python-multipart==0.0.9

cryptography==42.0.0

llama-cpp-python==0.2.90
huggingface-hub==0.24.0


# <-- !!! НАТИСНІТЬ ENTER ТУТ, ЩОБ БУВ ПОРОЖНІЙ РЯДОК !!!
